Lots of Comments adding & Cosmetics job, plus @todo for jobs yet to be done.
New class named "Drawable" to assemble Edge and Node(s) commons.
Everything in my commit turn arround the fact that lots of stuff could certainly be moved up in the class hierarchy, with some parameters to do the needed tweaking. Will post a "issue" to ask for permission to go farther.

Please excuse for the 2 class and the 5 or so functions moved : I know that it breaks the diffs and code history, but I think it worth it (better to do it at the very beginning)

git-svn-id: http://codeswarm.googlecode.com/svn/trunk@123 eda9c206-d64f-0410-a2f2-67f3fa0499ed