Changed to stream-based xml parsing: faster startup, reduced memory footprint
However, the event queue no longer sorts the values coming in, so it's the
user's responsibility to ensure that the input file is sorted.

Fortunately all of the methods that we've got for generating the xml files
seem to only produced sorted values.

The tradeoff that you get with this though seems worthwhile.  Loading a
50 meg input file used to take ~1GB of RAM and around 30 seconds of startup
time.  Now it uses ~200MB even after running for a long while, and starts
up almost immediately.