Added "PhysicsEngineOrderly" which runs in O(n) in the number of FileNodes
I created PhysicsEngineOrderly to deal with visualizations of large numbers of FileNodes on the screen at once.  It seems to make a pretty big difference for these kinds of visualizations.

It works by figuring out a place for each FileNode to go around its PersonNode just based on the number of FileNodes that are already around that PersonNode.  I had to add a field to PersonNode to keep track of which FileNodes are currently orbiting it, but since most visualizations, even large ones, have relatively few people involved, the memory overhead is minimal.  The visualization has less of an organic feel to it, I like the look of other physics engines more, but they just aren't always feasible.

One problem with the current visualization is that the FileNodes rearrange themselves as old members fade out, which creates a lot of distracting visual noise (though it is a bit cool looking, it conveys very little information).