Modified run.sh to (more) explicitly set classpath.
This may not be necessary, but it doesn't hurt. I'm having trouble
diagnosing a crash in Linux and I'm trying to track it down. I thought
this may fix it, but it doesn't -- but it doesn't seem to hurt, either,
so I'm going to leave it in in case it *is* necessary after all.