Simplified and refactored convert_logs.py
The goal of this was to have stream-based input and output so that
the initial startup of the visualization would be nearly instantaneous
even when a cache of the event xml wasn't present.

This was completely foiled.

The documentation for git log seems to
be false, as neither --reverse nor --date-order --reverse seem to
always give the commits in chronological order.

`svn log -r 1:HEAD` does give the commits in chronological order,
but this is useless because the svn server seems to take quite a
long time to start streaming back commits when this query is made.

Mercurial seems promising, as `hg log -r 1:` looks like it might work,
but the mercurial parser in convert_logs.py is broken.  The preferred
one to use is hg_log.py.

Memory usage has been cut down, though it was hardly a problem before
in this part.  If you squint your eyes and tilt your head it almost seems
like it's a little faster maybe.  I think it's cleaner and easier to modify
at least.