Initial experiment with adding gravatars for people
Pros: It's easier to remember and associate a face/identicon than a username/email address
      Looks neat

Cons: Currently it gets in the way of the visualization.
      Not all repositories customarily use emails for usernames

Making them a bit smaller, a bit more transparent, and widening the minimum radius that
file nodes circle at would go a long ways.  Obviously they'll be made optional, and we'll
probably expose many of the gravatar options.