I think there's something worthwhile here
The gravatars are now smaller, more centered, and fade out smoothly.

A bug with the positioning of nodes is currently the biggest problem I think.
I'm gonna go hack on that for the master branch.

Still no configuration, which is the next step I think.  I'd like the image
fetcher to be optional and configurable as well, to better accommodate
repositories where usernames aren't email addresses, but it still might be
possible to find a representative image for a user.