Fixed a long standing bug in PhysicsEngineOrderly's placement of FileNodes
I was only tracking the addition of FileNodes to the list of nodes that a
Person has edited, not the removal, which created that weird behavior of
nodes hanging out between people, and clustered over the center of a person.

Apparently I just forgot to do the second half.  Kinda embarassing, eheh.