Selectable AvatarFetcher and smarter loading of avatars to eliminate lag
Now supports the option "AvatarFetcher" which can be either GravatarFetcher or FreebaseAvatarFetcher.  Leave blank for no avatar images.  FreebaseAvatarFetcher is mostly useful for some visualizations I've been working on for visualizing freebase.com.

The avatar cache is primed as the xml is read in.  This, combined with a larger read ahead queue keeps lag pretty much nonexistent.