Added full support for darcs, including autodiscovery
Unfortunately darcs sometimes will produce invalid XML when asked.  This is a known bug, and related to character encoding issues that they seem unwilling to fix.  I might take the performance hit and rewrite this using BeautifulSoup at some point.  I've only seen the bug in the darcs repo itself, it comes up
when there are authors with non-ascii names.